refactor: extract mount path utilities to shared module
Move sanitizeMountSegment, repoBasename, adjustFileExtension,
getDefaultAgentMountPath, and getDefaultObjectMountPath to
src/utils/mount.ts. Both CLI (commands/devbox/create.ts) and TUI
(components/DevboxCreatePage.tsx) now import from the shared module.
